Modify rooks functions for efficiency
I realized that it would be more efficient to move the conflict check
immediately after every time a piece is placed.  Reason: if the
combination of piece placements up to a certain point has conflicts,
there's no point continuing to iterate through the remaining rows and
place pieces there.  As a result, the conflict check is moved; as well
as the output definition (i.e. solutionCount++), which now resides in
the if (currentRow === n) conditional statement (i.e., no more rows to
iterate through, thus we have a passing solution).

This makes the tests run much faster, avoiding more unnecessary checks.